lecture: 배열과 리스트, 제네릭과 무공변
