+++ inflearn-kotlin-advanced/src/main/kotlin/com/inflearn/kotlin/advanced/lec01/Cage.kt
@@ -1,5 +1,7 @@
 package com.inflearn.kotlin.advanced.lec01
 
+import com.inflearn.kotlin.advanced.lec02.Cage2
+
 class Cage {
     private val animals: MutableList<Animal> = mutableListOf()
 
@@ -10,20 +12,6 @@ class Cage {
     }
 
     fun moveFrom(cage: Cage) {
-        this.animals.addAll(cage.animals)
-    }
-}
-
-class Cage2<T> {
-    private val animals: MutableList<T> = mutableListOf()
-
-    fun getFirst(): T = animals.first()
-
-    fun put(animal: T) {
-        this.animals.add(animal)
-    }
-
-    fun moveFrom(cage: Cage2<T>) {
         this.animals.addAll(cage.animals)
     }
 }
@@ -40,10 +28,4 @@ fun main() {
     val cage2 = Cage2<Carp>()
     cage2.put(Carp("잉어"))
     val carp2 = cage2.getFirst()
-
-    val goldFishCage = Cage2<GoldFish>()
-    goldFishCage.put(GoldFish("금붕어"))
-
-    val fishCage = Cage2<Fish>()
-//    fishCage.moveFrom(goldFishCage)
 }
