--- inflearn-kotlin-advanced/src/main/kotlin/com/inflearn/kotlin/advanced/lec01/README.md
@@ -1,3 +1,0 @@
-# 1강. 제네릭과 타입 파라미터
-
-
